requirements.txt
alembic==1.18.3
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
greenlet==3.3.1
gunicorn==25.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
MouseInfo==0.1.3
openpyxl==3.1.5
packaging==26.0
psycopg2-binary==2.9.11
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.2.1
pytweening==1.2.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==2.3.0
WTForms==3.2.1
